Optional white theme
New custom var sayid-theme. Defaults to 'classic-dark. Which should keep everything the way it used to be.
When not 'classic-dark it will assign actual faces (as opposed to :foreground props).
Themes are defined under sayid-themes.

Rationale: The defaults are pretty opinionated. Unusable on my setup. There was no way to change that.