Fix dependency injection at jack-in time
Commit https://github.com/clojure-emacs/cider/commit/8989f40d in cider
stopped using `cider-jack-in-lein-plugins` as a method for defining
the dependencies to insert, which causes the sayid dependencies to be
missed upon jack-in.

Since we're using Cider anyways, and likely using it to define our
nREPL version, we're going to leverage the more generic
`cider-jack-in-dependencies` variable for defining the sayid
dependency.

To avoid potential version collisions, I've opted to no longer add the
sayid plugin to the `cider-jack-in-lein-plugins` variable.